# sdks/sdkwork-rtc-backend-sdk/openapi/sdkwork-rtc-backend-api.sdkgen.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork RTC Backend API",
    "version": "1.0.0",
    "description": "Backend/admin contract for SDKWork RTC rooms, provider profiles, provider routes, media sessions, media artifacts, provider webhooks, active provider query jobs, and quality samples.",
    "x-sdkwork-api-authority": "sdkwork-rtc-backend-api",
    "x-sdkwork-sdk-family": "sdkwork-rtc-backend-sdk",
    "x-sdkwork-owner": "sdkwork-rtc",
    "x-sdkwork-domain": "rtc"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18088",
      "description": "Local sdkwork-rtc runtime"
    }
  ],
...
  "security": [
    {
      "AuthToken": [],
      "AccessToken": []
    }
  ],
  "paths": {
...
    "/backend/v3/api/rtc/media_artifacts/{mediaArtifactId}": {
      "get": {
        "tags": [
          "rtcMediaArtifacts"
        ],
        "summary": "Rtc media Artifacts retrieve.",
        "operationId": "rtc.mediaArtifacts.retrieve",
        "parameters": [
          {
            "name": "mediaArtifactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RtcMediaArtifactResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-rtc",
        "x-sdkwork-api-authority": "sdkwork-rtc-backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-router-rtc-backend-api",
        "x-sdkwork-domain": "rtc",
        "x-sdkwork-resource": "rtc.mediaArtifacts",
        "x-sdkwork-request-context": "BackendRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "rtc.media_artifacts.read",
        "x-sdkwork-auth-mode": "dual-token"
      }
    },
...
    "/backend/v3/api/rtc/media_sessions/{mediaSessionId}": {
      "get": {
        "tags": [
          "rtcMediaSessions"
        ],
        "summary": "Rtc media Sessions retrieve.",
        "operationId": "rtc.mediaSessions.retrieve",
        "parameters": [
          {
            "name": "mediaSessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RtcMediaSessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-rtc",
        "x-sdkwork-api-authority": "sdkwork-rtc-backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-router-rtc-backend-api",
        "x-sdkwork-domain": "rtc",
        "x-sdkwork-resource": "rtc.mediaSessions",
        "x-sdkwork-request-context": "BackendRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "rtc.media_sessions.read",
        "x-sdkwork-auth-mode": "dual-token"
      }
    },
...
    "/backend/v3/api/rtc/media_sessions/{mediaSessionId}/completion_record": {
      "get": {
        "tags": [
          "rtcMediaSessions"
        ],
        "summary": "Rtc media Sessions completion Record retrieve.",
        "operationId": "rtc.mediaSessions.completionRecord.retrieve",
        "parameters": [
          {
            "name": "mediaSessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RtcMediaSessionCompletionRecordResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-rtc",
        "x-sdkwork-api-authority": "sdkwork-rtc-backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-router-rtc-backend-api",
        "x-sdkwork-domain": "rtc",
        "x-sdkwork-resource": "rtc.mediaSessions.completionRecord",
        "x-sdkwork-request-context": "BackendRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "rtc.media_sessions.read",
        "x-sdkwork-auth-mode": "dual-token"
      }
    },
...
    "/backend/v3/api/rtc/provider_accounts/{providerAccountId}": {
      "get": {
        "tags": [
          "rtcProviderAccounts"
        ],
        "summary": "Rtc provider Accounts retrieve.",
        "operationId": "rtc.providerAccounts.retrieve",
        "parameters": [
          {
            "name": "providerAccountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RtcProviderAccountResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-rtc",
        "x-sdkwork-api-authority": "sdkwork-rtc-backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-router-rtc-backend-api",
        "x-sdkwork-domain": "rtc",
        "x-sdkwork-resource": "rtc.providerAccounts",
        "x-sdkwork-request-context": "BackendRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "rtc.provider_accounts.read",
        "x-sdkwork-auth-mode": "dual-token"
      },
      "patch": {
        "tags": [
          "rtcProviderAccounts"
        ],
        "summary": "Rtc provider Accounts update.",
        "operationId": "rtc.providerAccounts.update",
        "parameters": [
          {
            "name": "providerAccountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RtcProviderAccountResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-rtc",
        "x-sdkwork-api-authority": "sdkwork-rtc-backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-router-rtc-backend-api",
        "x-sdkwork-domain": "rtc",
        "x-sdkwork-resource": "rtc.providerAccounts",
        "x-sdkwork-request-context": "BackendRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "rtc.provider_accounts.write",
        "x-sdkwork-auth-mode": "dual-token",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RtcProviderAccountCommand"
              }
            }
          }
        }
      }
    },
...
    "/backend/v3/api/rtc/provider_applications/{providerApplicationId}": {
      "get": {
        "tags": [
          "rtcProviderApplications"
        ],
        "summary": "Rtc provider Applications retrieve.",
        "operationId": "rtc.providerApplications.retrieve",
        "parameters": [
          {
            "name": "providerApplicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RtcProviderApplicationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-rtc",
        "x-sdkwork-api-authority": "sdkwork-rtc-backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-router-rtc-backend-api",
        "x-sdkwork-domain": "rtc",
        "x-sdkwork-resource": "rtc.providerApplications",
        "x-sdkwork-request-context": "BackendRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "rtc.provider_applications.read",
        "x-sdkwork-auth-mode": "dual-token"
      },
      "patch": {
        "tags": [
          "rtcProviderApplications"
        ],
        "summary": "Rtc provider Applications update.",
        "operationId": "rtc.providerApplications.update",
        "parameters": [
          {
            "name": "providerApplicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RtcProviderApplicationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-rtc",
        "x-sdkwork-api-authority": "sdkwork-rtc-backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-router-rtc-backend-api",
        "x-sdkwork-domain": "rtc",
        "x-sdkwork-resource": "rtc.providerApplications",
        "x-sdkwork-request-context": "BackendRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "rtc.provider_applications.write",
        "x-sdkwork-auth-mode": "dual-token",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RtcProviderApplicationCommand"
              }
            }
          }
        }
      }
    },
...
    "/backend/v3/api/rtc/provider_plugins/{provider}": {
      "get": {
        "tags": [
          "rtcProviderPlugins"
        ],
        "summary": "Rtc provider Plugins retrieve.",
        "operationId": "rtc.providerPlugins.retrieve",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RtcApiResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-rtc",
        "x-sdkwork-api-authority": "sdkwork-rtc-backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-router-rtc-backend-api",
        "x-sdkwork-domain": "rtc",
        "x-sdkwork-resource": "rtc.providerPlugins",
        "x-sdkwork-request-context": "BackendRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "rtc.provider_plugins.read",
        "x-sdkwork-auth-mode": "dual-token"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "SDKWork auth token carried as Authorization: Bearer <auth_token>."
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token",
        "description": "SDKWork access isolation token."
      }
    },
    "schemas": {
...
      "RtcMediaSessionResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "requestId",
          "data"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "data": {
            "$ref": "#/components/schemas/RtcMediaSession"
          }
        }
      },
...
      "RtcMediaArtifactResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "requestId",
          "data"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "data": {
            "$ref": "#/components/schemas/RtcMediaArtifact"
          }
        }
      },
...
      "RtcMediaSessionCompletionRecordResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "requestId",
          "data"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "data": {
            "$ref": "#/components/schemas/RtcMediaSessionCompletionRecord"
          }
        }
      },
...
      "RtcProviderAccountResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "requestId",
          "data"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "data": {
            "$ref": "#/components/schemas/RtcProviderAccount"
          }
        }
      },
...
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      }
    }
  },
...
}